move lock_if_required-decorator to datalad.utils
